[JDK17] Update working with PIDs to be compatible with JDK9+
Even though this change doesn't makes this project able to compile with
JDK9+ it at least adds a possibility to utilize its classes and methods
in some other projects with regards the work with PIDs running on JDK9+.

We are using a "hack" to call the method for PID retrieval by reflection
because it is only present in Java 9 and newer. That allows us to still
compile this code on older JDKs.

As a consequence the pid number changed from integer to long so we had
to change it everywhere. Dependent projects should update their usages
appropriately to long too, although since long -> integer is just about
the precision loss issue and I am not aware of any platform (from those
we use) that uses long instead of integer for their PIDs, it shouldn't
be a big issue.

+++ core/src/main/groovy/noe/common/newcmd/PsCmdFormat.java
@@ -27,11 +27,11 @@ public enum PsCmdFormat {
     /**
      * Id of process.
      */
-    PROCESS_ID(false, Integer.class),
+    PROCESS_ID(false, Long.class),
     /**
      * Pid of process parent.
      */
-    PARENT_PROCESS_ID(false, Integer.class),
+    PARENT_PROCESS_ID(false, Long.class),
     /**
      * Priority (nice) of the process.
      */
